New thread pool reactor implementation and refactoring world daemon. (#8)
* New thread pool reactor implementation and refactoring world daemon.

* Fix Linux build.

* Fix Windows debug builds.
