+++ src/mangosd/SOAP/SoapThread.cpp
@@ -22,39 +22,123 @@
  * and lore are copyrighted by Blizzard Entertainment, Inc.
  */
 
-#include "MaNGOSsoap.h"
+#include <ace/OS.h>
+#include <ace/Message_Block.h>
 
-#define POOL_SIZE   5
+#include "SoapThread.h"
+#include "soapH.h"
+#include "soapStub.h"
 
-void MaNGOSsoapRunnable::run()
+#include "World.h"
+#include "Log.h"
+#include "AccountMgr.h"
+
+
+/// WARNING! This code needs serious reviewing
+
+
+struct SOAPCommand
+{
+    public:
+        void appendToPrintBuffer(const char* msg)
+          { m_printBuffer += msg; }
+
+        void setCommandSuccess(bool val)
+         { m_success = val; }
+
+        bool hasCommandSucceeded()
+          { return m_success; }
+
+        static void print(void* callbackArg, const char* msg)
+        {
+            ((SOAPCommand*)callbackArg)->appendToPrintBuffer(msg);
+        }
+
+        static void commandFinished(void* callbackArg, bool success);
+
+        bool m_success;
+        std::string m_printBuffer;
+};
+
+
+class SoapPool : public ACE_Task<ACE_MT_SYNCH>
+{
+    public:
+        virtual int svc(void) override
+        {
+            while (1)
+            {
+                ACE_Message_Block* mb = 0;
+                if (this->getq(mb) == -1)
+                {
+                    break;
+                }
+
+                // Process the message.
+                process_message(mb);
+            }
+            return 0;
+        }
+
+    private:
+        void process_message(ACE_Message_Block* mb)
+        {
+            struct soap* soap;
+            ACE_OS::memcpy(&soap, mb->rd_ptr(), sizeof(struct soap*));
+            mb->release();
+
+            soap_serve(soap);
+
+            soap_destroy(soap); // dealloc C++ data
+            soap_end(soap); // dealloc data and clean up
+            soap_done(soap); // detach soap struct
+            free(soap);
+        }
+};
+
+
+
+SoapThread::~SoapThread()
+{
+  if(pool_)
+      delete pool_;
+}
+
+int SoapThread::open(void* unused)
 {
     // create pool
-    SOAPWorkingThread pool;
-    pool.activate(THR_NEW_LWP | THR_JOINABLE, POOL_SIZE);
+    pool_ = new SoapPool;
+    pool_->activate(THR_NEW_LWP | THR_JOINABLE, SOAP_THREADS);
 
-    struct soap soap;
-    int m, s;
-    soap_init(&soap);
-    soap_set_imode(&soap, SOAP_C_UTFSTRING);
-    soap_set_omode(&soap, SOAP_C_UTFSTRING);
-    m = soap_bind(&soap, m_host.c_str(), m_port, 100);
+    int m;
+    soap_init(&soap_);
+    soap_set_imode(&soap_, SOAP_C_UTFSTRING);
+    soap_set_omode(&soap_, SOAP_C_UTFSTRING);
+    m = soap_bind(&soap_, host_, port_, 100);
 
-    // check every 3 seconds if world ended
-    soap.accept_timeout = 3;
-
-    soap.recv_timeout = 5;
-    soap.send_timeout = 5;
     if (m < 0)
     {
-        sLog.outError("MaNGOSsoap: couldn't bind to %s:%d", m_host.c_str(), m_port);
-        exit(-1);
+        sLog.outError("SoapThread: couldn't bind to %s:%d", host_, port_);
+        return -1;
     }
 
-    sLog.outString("MaNGOSsoap: bound to http://%s:%d", m_host.c_str(), m_port);
+    // check every 3 seconds if world ended
+    soap_.accept_timeout = 3;
+    soap_.recv_timeout = 5;
+    soap_.send_timeout = 5;
 
+    activate();
+    return 0;
+}
+
+
+int SoapThread::svc()
+{
+    int s;
+    sLog.outString("SOAP Thread started (listening on %s:%d)", host_, port_);
     while (!World::IsStopped())
     {
-        s = soap_accept(&soap);
+        s = soap_accept(&soap_);
 
         if (s < 0)
         {
@@ -62,38 +146,28 @@ void MaNGOSsoapRunnable::run()
             continue;
         }
 
-        DEBUG_LOG("MaNGOSsoap: accepted connection from IP=%d.%d.%d.%d", (int)(soap.ip >> 24) & 0xFF, (int)(soap.ip >> 16) & 0xFF, (int)(soap.ip >> 8) & 0xFF, (int)soap.ip & 0xFF);
-        struct soap* thread_soap = soap_copy(&soap);// make a safe copy
+        struct soap* thread_soap = soap_copy(&soap_);// make a safe copy
 
         ACE_Message_Block* mb = new ACE_Message_Block(sizeof(struct soap*));
         ACE_OS::memcpy(mb->wr_ptr(), &thread_soap, sizeof(struct soap*));
-        pool.putq(mb);
+        pool_->putq(mb);
     }
-    pool.msg_queue()->deactivate();
-    pool.wait();
 
-    soap_done(&soap);
+    pool_->msg_queue()->deactivate();
+    pool_->wait();
+
+    soap_done(&soap_);
+    sLog.outString("SOAP Thread stopped");
+    return 0;
 }
 
-void SOAPWorkingThread::process_message(ACE_Message_Block* mb)
-{
-    ACE_TRACE(ACE_TEXT("SOAPWorkingThread::process_message"));
 
-    struct soap* soap;
-    ACE_OS::memcpy(&soap, mb->rd_ptr(), sizeof(struct soap*));
-    mb->release();
-
-    soap_serve(soap);
-    soap_destroy(soap); // dealloc C++ data
-    soap_end(soap); // dealloc data and clean up
-    soap_done(soap); // detach soap struct
-    free(soap);
-}
 /*
 Code used for generating stubs:
 
 int ns1__executeCommand(char* command, char** result);
 */
+
 int ns1__executeCommand(soap* soap, char* command, char** result)
 {
     // security check
@@ -135,15 +209,7 @@ int ns1__executeCommand(soap* soap, char* command, char** result)
         sWorld.QueueCliCommand(cmd);
     }
 
-    // wait for callback to complete command
-
-    int acc = connection.pendingCommands.acquire();
-    if (acc)
-    {
-        sLog.outError("MaNGOSsoap: Error while acquiring lock, acc = %i, errno = %u", acc, errno);
-    }
-
-    // alright, command finished
+    ACE_OS::sleep(1);
 
     char* printBuffer = soap_strdup(soap, connection.m_printBuffer.c_str());
     if (connection.hasCommandSucceeded())
@@ -159,7 +225,6 @@ void SOAPCommand::commandFinished(void* soapconnection, bool success)
 {
     SOAPCommand* con = (SOAPCommand*)soapconnection;
     con->setCommandSuccess(success);
-    con->pendingCommands.release();
 }
 
 ////////////////////////////////////////////////////////////////////////////////
