+++ src/mangosd/SOAP/SoapThread.h
@@ -1,0 +1,53 @@
+/**
+ * MaNGOS is a full featured server for World of Warcraft, supporting
+ * the following clients: 1.12.x, 2.4.3, 3.3.5a, 4.3.4a and 5.4.8
+ *
+ * Copyright (C) 2005-2017  MaNGOS project <https://getmangos.eu>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ *
+ * World of Warcraft, and all World of Warcraft or Warcraft art, images,
+ * and lore are copyrighted by Blizzard Entertainment, Inc.
+ */
+
+#ifndef MANGOS_H_SOAPTHREAD
+#define MANGOS_H_SOAPTHREAD
+
+#include <ace/Task.h>
+
+#include "Common.h"
+#include "soapH.h"
+
+class SoapPool;
+
+class SoapThread: public ACE_Task_Base
+{
+    enum { SOAP_THREADS = 1 }; //TODO: configure in mangosd.conf
+
+    public:
+        SoapThread(uint16 port, const char* host) : host_(host), port_(port), pool_(NULL) {}
+        virtual ~SoapThread();
+
+        virtual int svc() override;
+        virtual int open(void*) override;
+
+    private:
+        const char   *host_;
+        uint16       port_;
+        SoapPool     *pool_;
+        struct soap  soap_;
+};
+
+#endif
